src/main/java/br/filipov/values/c/CurvaAberturaFechamentoValues.java
package br.filipov.values.c;

/**
 * @author Marcelo A. Filipov
 * @since 2014
 */
public enum CurvaAberturaFechamentoValues {

	ABERTA(1, "Abertura"), FECHADA(2, "Fechamento");

	private Integer codigo;
	private String descricao;

	private CurvaAberturaFechamentoValues(Integer codigo, String descricao) {
		this.codigo = codigo;
		this.descricao = descricao;
	}

	public static String getDescricaoById(byte id) {
		String descricao = null;
		for (CurvaAberturaFechamentoValues v : values()) {
			if (v.getCodigo() == id) {
				descricao = v.getDescricao();
				break;
			}
		}
		return descricao;
	}

	public static String getDescricaoById(short id) {
		String descricao = null;
		for (CurvaAberturaFechamentoValues v : values()) {
			if (v.getCodigo() == id) {
				descricao = v.getDescricao();
				break;
			}
		}
		return descricao;
	}

	public static String getDescricaoById(int id) {
		String descricao = null;
		for (CurvaAberturaFechamentoValues v : values()) {
			if (v.getCodigo() == id) {
				descricao = v.getDescricao();
				break;
			}
		}
		return descricao;
	}
	
	public static String getDescricaoById(long id) {
		String descricao = null;
		for (CurvaAberturaFechamentoValues v : values()) {
			if (v.getCodigo() == id) {
				descricao = v.getDescricao();
				break;
			}
		}
		return descricao;
	}
	
	public static CurvaAberturaFechamentoValues parse(int id){
		CurvaAberturaFechamentoValues retorno = null;
		for (CurvaAberturaFechamentoValues v: values()){
			if (v.getCodigo() == id) {
				retorno = v;
				break;
			}
		}
		return retorno;
	}

	public Integer getCodigo() {
		return codigo;
	}

	public String getDescricao() {
		return descricao;
	}

}
